# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoneyFlex API",
        "description": "A REST API over customers and the services they hold. Each customer may\nhave many services.\n\n**Authentication**\n\nEvery endpoint except the health check and login requires credentials.\nTwo schemes are accepted and either is sufficient:\n\n- `basicAuth` — HTTP Basic, the scheme the API is required to support.\n- `bearerAuth` — a JWT obtained from `POST /api/v1/auth/login`.\n\nUse the **Authorize** button above to supply either one. The seeded\naccount is `api@moneyflex.test` / `password123`.",
        "contact": {
            "name": "MoneyFlex"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "View all services",
                "description": "A paginated listing across every customer. Responses are cached and invalidated whenever a service changes.",
                "operationId": "listServices",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Matches name or description.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "suspended",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Restrict to one customer.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Column to sort by. Prefix with \"-\" for descending.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "price",
                                "status",
                                "starts_at",
                                "ends_at",
                                "created_at",
                                "updated_at",
                                "-id",
                                "-name",
                                "-price",
                                "-status",
                                "-starts_at",
                                "-ends_at",
                                "-created_at",
                                "-updated_at"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Between 1 and 100.",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page of services.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Service"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/services/{service}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "View a service",
                "description": "Includes the owning customer.",
                "operationId": "showService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The service, with its customer embedded.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/Service"
                                                },
                                                {
                                                    "properties": {
                                                        "customer": {
                                                            "$ref": "#/components/schemas/Customer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update a service",
                "description": "Every field is optional. PATCH behaves identically. The owning customer cannot be changed here.",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Payment Gateway",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "At most two decimal places.",
                                        "type": "string",
                                        "example": "299.00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "suspended",
                                        "enum": [
                                            "active",
                                            "suspended",
                                            "cancelled"
                                        ]
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "ends_at": {
                                        "description": "Must not be earlier than starts_at.",
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated service.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service",
                "description": "A soft delete. The record stops appearing in listings and reads.",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted. No content is returned."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "title": "Customer",
                "properties": {
                    "id": {
                        "description": "The document level OpenAPI definition: metadata, servers, the two accepted\nsecurity schemes, and the schemas shared across endpoints.\n\nThis class holds no behaviour. It exists so the specification is written in\none readable place instead of being scattered across controllers.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Holdings"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "acme@moneyflex.test"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+971500000001",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "1 Sheikh Zayed Road, Dubai",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "services_count": {
                        "description": "Present on listings and single reads.",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Service": {
                "title": "Service",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Payment Gateway"
                    },
                    "description": {
                        "type": "string",
                        "example": "Card acquiring and settlement",
                        "nullable": true
                    },
                    "price": {
                        "description": "A decimal carried as a string so it never loses precision in transit.",
                        "type": "string",
                        "example": "249.99"
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "suspended",
                            "cancelled"
                        ]
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01",
                        "nullable": true
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2027-01-01",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 4
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 52
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination links",
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Validation error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The email field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The requested resource was not found."
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "Unauthorized": {
                "description": "No credentials were supplied, or they were not accepted.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Unauthenticated."
                        }
                    }
                }
            },
            "NotFound": {
                "description": "No such record.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ValidationFailed": {
                "description": "The request body or query string did not validate.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic Authentication using the account email and password.",
                "scheme": "basic"
            },
            "bearerAuth": {
                "type": "http",
                "description": "A JWT issued by POST /auth/login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}